Implement article's software-based DCSS and DCAS
The article explains how it works, I have the info for it commented above
I have not yet tested this and it likely breaks.  Managing the descriptor pointers is difficult as working with generic pointers in C++ became overly complex, and so I combined the descriptor with the mound node struct, since they get reused together anyway.

I trust the article that the DCAS implementation is lock-free.  Basically it stores the info for the operation in a "descriptor" that gets put in shared memory (the atomic being accessed).  If another thread wishes to alter the atomic but sees the descriptor, it will help finish the previous operation before performing its update.  Thus no threads lock each other as they will all still be performing work.